MMCA 4770 : Upgrade scala & library dependancies to latest version (#128)
* MMCA-4770 : initial commit

* MMCA-4770 : added unused warning, fixed ErrorHandler, Layout

* MMCA-4770 : initial fixes for test cases

* MMCA-4770 : test cases fixes - httpclient and mock changes

* MMCA-4770 : fixed url formation

* MMCA-4770 : fixed mock of http call in connectorSpec

* MMCA-4770 : fixed scalastyle issues

* MMCA-4770 :  compilation fixes in html view component

* MMCA-4770 : revoked changes of ErrorHandler to use LegacyFrontendErrorHandler

* MMCA-4770: PR comment fixes 01

+++ app/repositories/CacheRepository.scala
@@ -32,6 +32,9 @@ import java.util.concurrent.TimeUnit
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
 
+import org.mongodb.scala.SingleObservableFuture
+import org.mongodb.scala.ToSingleObservablePublisher
+
 @Singleton
 class DefaultCacheRepository @Inject()(mongo: MongoComponent,
                                        config: Configuration,
@@ -45,7 +48,7 @@ class DefaultCacheRepository @Inject()(mongo: MongoComponent,
       IndexModel(
         ascending("lastUpdated"),
         IndexOptions().name("cash-account-cache-last-updated-index")
-          .expireAfter(config.get[Int]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
+          .expireAfter(config.get[Long]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
       )
     )) with CacheRepository {
